Changed Search conversion event to trigger when Markdown has all been scrolled through viewport
