+++ apps/devportal/src/pages/_app.tsx
@@ -7,7 +7,7 @@ import Navbar from '@src/components/navigation/NavBar';
 import SearchInputSwitcher from '@src/components/sitecore-search/SearchInputSwitcher';
 import { AppProps } from 'next/app';
 import { Router } from 'next/router';
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import TagManager from 'react-gtm-module';
 import TopBarProgress from 'react-topbar-progress-indicator';
 import { AvenirNextR } from 'ui/common/fonts/avenirNextR';
@@ -23,21 +23,6 @@ function MyApp({ Component, pageProps }: AppProps) {
     },
     shadowBlur: 2,
   });
-
-  const onScroll = useCallback(() => {
-    const params = new URLSearchParams(window.location.search); // id=123
-    const fromSearch = params.get('fromSearch');
-    if (fromSearch && window.innerHeight + window.scrollY >= document.body.offsetHeight) {
-      trackEntityPageViewEvent('content', {
-        items: [
-          {
-            id: process.env.NEXT_PUBLIC_SEARCH_DOMAIN_ID_PREFIX + document.location.pathname.replace(/[/:.]/g, '_').replace(/_+$/, ''),
-          },
-        ],
-        actionSubtype: 'conversion',
-      });
-    }
-  }, []);
 
   Router.events.on('routeChangeStart', () => {
     setProgress(true);
@@ -60,11 +45,6 @@ function MyApp({ Component, pageProps }: AppProps) {
       PageController.getContext().setLocale({ country: 'us', language: 'en' });
       trackEntityPageViewEvent('content', { items: [{ id: process.env.NEXT_PUBLIC_SEARCH_DOMAIN_ID_PREFIX + document.location.pathname.replace(/[/:.]/g, '_').replace(/_+$/, '') }] });
     }
-
-    window.addEventListener('scroll', onScroll, { passive: true });
-    return () => {
-      window.removeEventListener('scroll', onScroll);
-    };
   });
 
   const SearchWrapper = ({ children }: any) => (IsSearchEnabled() ? <WidgetsProvider {...SEARCH_CONFIG}>{children}</WidgetsProvider> : children);
